requirements.txt
# AI Marketing Automation Platform
# Python 3.9+

# Video (Kling JWT + 长视频拼接 + 后制 TTS)
PyJWT>=2.8.0
moviepy>=1.0.3
edge-tts>=6.1.0

# Data collection
requests>=2.31.0
httpx>=0.25.0

# ETL
pandas>=2.0.0
python-dotenv>=1.0.0

# Utils
pydantic>=2.0.0

# AI / LLM
openai>=1.0.0
dashscope>=1.14.0

# Dashboard
streamlit>=1.28.0

# Scheduler
schedule>=1.2.0

# RAG / Vector Store
chromadb>=0.4.0

# Optional: for OpenAI embeddings in RAG
# tiktoken>=0.5.0
